+++ slsReceiverSoftware/src/Receiver.cpp
@@ -129,7 +129,7 @@ int64_t Receiver::getReceiverVersion() {
 }
 
 void Receiver::registerCallBackStartAcquisition(
-    int (*func)(std::string, std::string, uint64_t, uint32_t, void *),
+    int (*func)(std::string, std::string, uint64_t, size_t, void *),
     void *arg) {
     tcpipInterface->registerCallBackStartAcquisition(func, arg);
 }
@@ -141,13 +141,12 @@ void Receiver::registerCallBackAcquisitionFinished(void (*func)(uint64_t,
 }
 
 void Receiver::registerCallBackRawDataReady(
-    void (*func)(sls_receiver_header *, char *, uint32_t, void *), void *arg) {
+    void (*func)(sls_receiver_header *, char *, size_t, void *), void *arg) {
     tcpipInterface->registerCallBackRawDataReady(func, arg);
 }
 
 void Receiver::registerCallBackRawDataModifyReady(
-    void (*func)(sls_receiver_header *, char *, uint32_t &, void *),
-    void *arg) {
+    void (*func)(sls_receiver_header *, char *, size_t &, void *), void *arg) {
     tcpipInterface->registerCallBackRawDataModifyReady(func, arg);
 }
